requirements.txt
matplotlib==3.7.2
scikit-learn==1.2.2
keras==2.11.0
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-learn>=0.24.0
imbalanced-learn>=0.8.0